requirements.txt
yt-dlp>=2023.12.30
ffmpeg-python>=0.2.0
typing>=3.7.4
youtube_transcript_api==1.2.2
requests==2.32.3
openai>=1.30.0
python-dotenv>=1.0.1
python-dotenv==1.1.1
selenium==4.35.0